@@ -1,16 +1,18 @@
-require File.expand_path("../boot", __FILE__)
+# frozen_string_literal: true
 
-require "rails"
+require File.expand_path('boot', __dir__)
+
+require 'rails'
 # Pick the frameworks you want:
-require "active_model/railtie"
-require "active_job/railtie"
-require "active_record/railtie"
-require "action_controller/railtie"
-require "action_mailer/railtie"
-require "action_view/railtie"
-require "action_cable/engine"
+require 'active_model/railtie'
+require 'active_job/railtie'
+require 'active_record/railtie'
+require 'action_controller/railtie'
+require 'action_mailer/railtie'
+require 'action_view/railtie'
+require 'action_cable/engine'
 # require "sprockets/railtie"
-require "rails/test_unit/railtie"
+require 'rails/test_unit/railtie'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -30,7 +32,7 @@ module CashFlow
     # don't generate RSpec tests for views and helpers
     config.generators do |g|
       g.test_framework :rspec, fixture: true
-      g.fixture_replacement :factory_girl, dir: "spec/factories"
+      g.fixture_replacement :factory_girl, dir: 'spec/factories'
       g.view_specs false
       g.helper_specs false
       g.stylesheets = false
@@ -38,14 +40,13 @@ module CashFlow
       g.helper = false
     end
 
-    config.autoload_paths += %W(\#{config.root}/lib)
+    config.autoload_paths += %W[\#{config.root}/lib]
 
     config.middleware.insert_before 0, Rack::Cors do
       allow do
         origins '*'
-        resource '*', :headers => :any, :methods => [:get, :post, :put, :delete, :options]
+        resource '*', headers: :any, methods: %i[get post put delete options]
       end
     end
-
   end
 end
