@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Api
   module V1
     module Concerns
@@ -11,7 +13,7 @@ module Api
         # Devise methods overwrites
         def current_user
           @current_user ||=
-            User.find_by(auth_token: request.headers["Authorization"])
+            User.find_by(auth_token: request.headers['Authorization'])
         end
 
         def user_signed_in?
