@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Api
   module V1
     module Concerns
@@ -16,7 +18,7 @@ module Api
         private
 
         def extract_locale_from_accept_language_header
-          req_lang = request.env["HTTP_ACCEPT_LANGUAGE"]
+          req_lang = request.env['HTTP_ACCEPT_LANGUAGE']
           if req_lang
             language = req_lang.scan(/^[a-z]{2}/).first
             language if I18n.available_locales.include?(language.to_sym)
