@@ -1,35 +1,37 @@
+# frozen_string_literal: true
+
 module Api
   module V1
     class DebtGroupsController < ApiController
-      skip_before_action :auth_with_token!, only: [:create, :data, :all, :update, :destroy]
+      skip_before_action :auth_with_token!, only: %i[create data all update destroy]
 
       def data
-        account = current_user.accounts.first
-        groups = DebtGroup.where(account_id: account.id).order("name ASC").pluck(:name)
+        account = @current_user.accounts.first
+        groups = DebtGroup.where(account_id: account.id).order('name ASC').pluck(:name)
 
         render json: {
-                 status: "SUCCESS",
-                 message: "Loaded user debt groups",
-                 groups: groups,
-               }, status: :ok
+          status: 'SUCCESS',
+          message: 'Loaded user debt groups',
+          groups: groups
+        }, status: :ok
       end
 
       def all
-        account = current_user.accounts.first
+        account = @current_user.accounts.first
         groups = DebtGroup.where(account_id: account.id)
 
         render json: {
-                 status: "SUCCESS",
-                 message: "Loaded user debt groups",
-                 groups: groups,
-               }, status: :ok
+          status: 'SUCCESS',
+          message: 'Loaded user debt groups',
+          groups: groups
+        }, status: :ok
       end
 
       def create
-        new_group = DebtGroup.new()
-        new_group.account_id = Account.where(user_id: User.where(auth_token: params["headers"]["Authorization"]).first.id).first.id
-        new_group.name = params["params"]["name"]
-        new_group.description = params["params"]["description"]
+        new_group = DebtGroup.new
+        new_group.account_id = Account.where(user_id: User.where(auth_token: params['headers']['Authorization']).first.id).first.id
+        new_group.name = params['params']['name']
+        new_group.description = params['params']['description']
 
         if new_group.save
           render json: new_group, status: :created
@@ -39,37 +41,37 @@ module Api
       end
 
       def update
-        group = DebtGroup.find(Integer(params["params"]["id"]))
-        name = params["params"]["name"]
-        description = params["params"]["description"]
+        group = DebtGroup.find(Integer(params['params']['id']))
+        name = params['params']['name']
+        description = params['params']['description']
 
         group.update(name: name, description: description)
 
         if group.save
           render json: {
-            status: "SUCCESS",
-            message: "Debt Group updated",
+            status: 'SUCCESS',
+            message: 'Debt Group updated'
           }, status: :ok
         else
           render json: {
-            status: "ERROR",
-            message: "update error",
+            status: 'ERROR',
+            message: 'update error'
           }, status: :unprocessible_entity
         end
       end
 
       def destroy
-        group = DebtGroup.destroy(params["id"])
+        group = DebtGroup.destroy(params['id'])
 
         if group
           render json: {
-            status: "SUCCESS",
-            message: "Debt Group deleted",
+            status: 'SUCCESS',
+            message: 'Debt Group deleted'
           }, status: :ok
         else
           render json: {
-            status: "ERROR",
-            message: "Invalid id",
+            status: 'ERROR',
+            message: 'Invalid id'
           }, status: 400
         end
       end
