@@ -1,37 +1,39 @@
+# frozen_string_literal: true
+
 module Api
   module V1
     class ExpenseGroupsController < ApiController
       # before_action :set_expense, only: [:show, :edit, :update, :destroy]
-      skip_before_action :auth_with_token!, only: [:create, :data, :all, :update, :destroy]
+      skip_before_action :auth_with_token!, only: %i[create data all update destroy]
 
       # GET /expense_groups
       def data
-        account = current_user.accounts.first
-        expense_groups = ExpenseGroup.where(account_id: account.id).order("name ASC").pluck(:name)
+        account = @current_user.accounts.first
+        expense_groups = ExpenseGroup.where(account_id: account.id).order('name ASC').pluck(:name)
 
         render json: {
-                 status: "SUCCESS",
-                 message: "Loaded user expense groups",
-                 expense_groups: expense_groups,
-               }, status: :ok
+          status: 'SUCCESS',
+          message: 'Loaded user expense groups',
+          expense_groups: expense_groups
+        }, status: :ok
       end
 
       def all
-        account = current_user.accounts.first
+        account = @current_user.accounts.first
         expense_groups = ExpenseGroup.where(account_id: account.id)
 
         render json: {
-                 status: "SUCCESS",
-                 message: "Loaded user expense groups",
-                 expense_groups: expense_groups,
-               }, status: :ok
+          status: 'SUCCESS',
+          message: 'Loaded user expense groups',
+          expense_groups: expense_groups
+        }, status: :ok
       end
 
       def create
-        new_group = ExpenseGroup.new()
-        new_group.account_id = Account.where(user_id: User.where(auth_token: params["headers"]["Authorization"]).first.id).first.id
-        new_group.name = params["params"]["name"]
-        new_group.description = params["params"]["description"]
+        new_group = ExpenseGroup.new
+        new_group.account_id = Account.where(user_id: User.where(auth_token: params['headers']['Authorization']).first.id).first.id
+        new_group.name = params['params']['name']
+        new_group.description = params['params']['description']
 
         if new_group.save
           render json: new_group, status: :created
@@ -41,37 +43,37 @@ module Api
       end
 
       def update
-        group = ExpenseGroup.find(Integer(params["params"]["id"]))
-        name = params["params"]["name"]
-        description = params["params"]["description"]
+        group = ExpenseGroup.find(Integer(params['params']['id']))
+        name = params['params']['name']
+        description = params['params']['description']
 
         group.update(name: name, description: description)
 
         if group.save
           render json: {
-            status: "SUCCESS",
-            message: "Expense Group updated",
+            status: 'SUCCESS',
+            message: 'Expense Group updated'
           }, status: :ok
         else
           render json: {
-            status: "ERROR",
-            message: "update error",
+            status: 'ERROR',
+            message: 'update error'
           }, status: :unprocessible_entity
         end
       end
 
       def destroy
-        group = ExpenseGroup.destroy(params["id"])
+        group = ExpenseGroup.destroy(params['id'])
 
         if group
           render json: {
-            status: "SUCCESS",
-            message: "Expense Group deleted",
+            status: 'SUCCESS',
+            message: 'Expense Group deleted'
           }, status: :ok
         else
           render json: {
-            status: "ERROR",
-            message: "Invalid id",
+            status: 'ERROR',
+            message: 'Invalid id'
           }, status: 400
         end
       end
