@@ -1,4 +1,6 @@
-require "spec_helper"
+# frozen_string_literal: true
+
+require 'spec_helper'
 
 module Api
   module V1
@@ -10,42 +12,42 @@ module Api
       end
 
       before do
-        routes.draw { get "fake_not_found" => "api/v1/api#fake_not_found" }
+        routes.draw { get 'fake_not_found' => 'api/v1/api#fake_not_found' }
       end
 
-      it "return english message when no locale" do
+      it 'return english message when no locale' do
         signed_get :fake_not_found, params: {}
-        expect(json_response["error"]["message"]).to eq I18n.t(
-             "errors.messages.not_found", locale: I18n.default_locale,
-           )
-        expect(json_response["error"]["status"]).to eq 404
+        expect(json_response['error']['message']).to eq I18n.t(
+          'errors.messages.not_found', locale: I18n.default_locale
+        )
+        expect(json_response['error']['status']).to eq 404
       end
 
-      it "return english message when en locale" do
-        request.env["HTTP_ACCEPT_LANGUAGE"] = "en"
+      it 'return english message when en locale' do
+        request.env['HTTP_ACCEPT_LANGUAGE'] = 'en'
         signed_get :fake_not_found, params: {}
-        expect(json_response["error"]["message"]).to eq I18n.t(
-             "errors.messages.not_found", locale: :en,
-           )
-        expect(json_response["error"]["status"]).to eq 404
+        expect(json_response['error']['message']).to eq I18n.t(
+          'errors.messages.not_found', locale: :en
+        )
+        expect(json_response['error']['status']).to eq 404
       end
 
-      it "return spanish message when es locale" do
-        request.env["HTTP_ACCEPT_LANGUAGE"] = "es"
+      it 'return spanish message when es locale' do
+        request.env['HTTP_ACCEPT_LANGUAGE'] = 'es'
         signed_get :fake_not_found, params: {}
-        expect(json_response["error"]["message"]).to eq I18n.t(
-             "errors.messages.not_found", locale: :es,
-           )
-        expect(json_response["error"]["status"]).to eq 404
+        expect(json_response['error']['message']).to eq I18n.t(
+          'errors.messages.not_found', locale: :es
+        )
+        expect(json_response['error']['status']).to eq 404
       end
 
-      it "return english message when unknown locale" do
-        request.env["HTTP_ACCEPT_LANGUAGE"] = "wk"
+      it 'return english message when unknown locale' do
+        request.env['HTTP_ACCEPT_LANGUAGE'] = 'wk'
         signed_get :fake_not_found, params: {}
-        expect(json_response["error"]["message"]).to eq I18n.t(
-             "errors.messages.not_found", locale: I18n.default_locale,
-           )
-        expect(json_response["error"]["status"]).to eq 404
+        expect(json_response['error']['message']).to eq I18n.t(
+          'errors.messages.not_found', locale: I18n.default_locale
+        )
+        expect(json_response['error']['status']).to eq 404
       end
     end
   end
