@@ -1,7 +1,9 @@
+# frozen_string_literal: true
+
 FactoryGirl.define do
   factory :note do
-    title "MyString"
-    content "MyText"
+    title 'MyString'
+    content 'MyText'
     user
   end
 end
